# assets/openapi.json
{
  "openapi": "3.1.1",
  "info": {
    "version": "1.0.0",
    "title": "Takealot Marketplace API",
    "summary": "A service for merchants to integrate with Takealot seller information.",
    "description": "This API allows merchants to manage their seller information, offers, orders, returns, shipments, and more on the Takealot marketplace platform."
  },
  "servers": [
    {
      "url": "https://marketplace-api.takealot.com/v1",
      "description": "Production API"
    },
    {
      "url": "http://localhost:8000/v1",
      "description": "Local dev and functional testing API"
    }
  ],
  "security": [
    {
      "api_key": []
    }
  ],
...
  "components": {
    "securitySchemes": {
      "api_key": {
        "type": "apiKey",
        "description": "The API key",
        "name": "X-API-Key",
        "in": "header"
      }
    },
    "schemas": {
      "OfferID": {
        "type": "integer",
        "description": "Unique identifier for the offer. Unique across all offers.",
        "example": 123456,
        "readOnly": true,
        "format": "int32",
        "minimum": 1
      },
      "DefaultError": {
        "type": "object",
        "description": "A default error response containing an error message.",
        "required": [
          "message"
        ],
        "properties": {
          "message": {
            "type": "string",
            "description": "A human-readable message describing the error.",
            "example": "An unexpected error occurred."
          }
        },
        "additionalProperties": false
      },
...
      "Count": {
        "type": "integer",
        "description": "Total count of items available with the given filters.",
        "example": 150
      },
      "Limit": {
        "type": "integer",
        "description": "The maximum number of items returned at a time in the response.",
        "example": 100
      },
      "ContinuationToken": {
        "type": "string",
        "description": "A token to retrieve the next page of results. If not provided, there are no more results.",
        "example": "eyJvZmZzZXQiOiAxMDB9"
      },
      "ForbiddenError": {
        "type": "object",
        "description": "An error response returned when the request is forbidden, typically due to insufficient permissions or invalid authentication.",
        "properties": {
          "errors": {
            "type": "array",
            "description": "List of errors explaining why the request was forbidden.",
            "items": {
              "type": "object",
              "description": "Details of a specific error.",
              "properties": {
                "status": {
                  "type": "integer",
                  "description": "The HTTP status code for the error.",
                  "example": 403
                },
                "title": {
                  "type": "string",
                  "description": "A brief description of the error.",
                  "example": "Forbidden"
                },
                "type": {
                  "type": "string",
                  "description": "A reference identifying the error type."
                }
              }
            }
          }
        }
      },
      "ValidationError": {
        "type": "object",
        "description": "An error response returned when request validation fails, containing details about the validation errors.",
        "properties": {
          "errors": {
            "type": "array",
            "description": "List of validation errors that occurred.",
            "items": {
              "type": "object",
              "description": "Details of a specific validation error.",
              "properties": {
                "status": {
                  "type": "integer",
                  "description": "The HTTP status code for the error.",
                  "example": 400
                },
                "title": {
                  "type": "string",
                  "description": "A brief description of the validation error.",
                  "example": "Invalid field value"
                },
                "type": {
                  "type": "string",
                  "description": "A URI reference identifying the error type.",
                  "example": "https://api.takealot.com/errors/validation"
                }
              }
            }
          }
        }
      },
      "Error": {
        "type": "object",
        "description": "A simple error object containing an error title.",
        "properties": {
          "title": {
            "type": "string",
            "description": "A brief description of the error.",
            "example": "Resource not found"
          }
        }
      },
      "SKURequestBody": {
        "type": "string",
        "minLength": 1,
        "maxLength": 255,
        "pattern": "^[a-zA-Z0-9-_/.]+$",
        "description": "SKU (Stock Keeping Unit) for the offer (provided by the merchant)",
        "example": "SKU-ABC123"
      },
      "SellingPriceRequestBody": {
        "type": "integer",
        "format": "int32",
        "minimum": 1,
        "description": "Selling Price in South African Rand.",
        "example": 345
      },
      "RRPRequestBody": {
        "type": "integer",
        "format": "int32",
        "minimum": 1,
        "description": "RRP (Recommended Retail Price) in South African Rand.",
        "example": 399
      },
...
    },
    "parameters": {
      "OfferID": {
        "name": "offer_id",
        "description": "Takealot offer ID",
        "in": "query",
        "required": false,
        "example": 1234567890,
        "schema": {
          "$ref": "#/components/schemas/OfferID"
        }
      },
      "OfferIDIn": {
        "description": "Filter by offers with these IDs.",
        "name": "offer_id__in",
        "in": "query",
        "required": false,
        "example": [
          1234567,
          1234568,
          1234569
        ],
        "schema": {
          "type": "array",
          "items": {
            "$ref": "#/components/schemas/OfferID"
          }
        }
      },
      "OfferFields": {
        "description": "Return only these fields.",
        "name": "fields",
        "in": "query",
        "required": false,
        "example": [
          "offer_id",
          "sku",
          "selling_price",
          "status",
          "title"
        ],
        "schema": {
          "type": "array",
          "items": {
            "type": "string",
            "enum": [
              "offer_id",
              "tsin_id",
              "sku",
              "barcode",
              "product_label",
              "selling_price",
              "rrp",
              "minimum_leadtime_days",
              "status",
              "title",
              "discount_percentage",
              "storage_fee_eligible",
              "created_at",
              "updated_at",
              "affected_by_vacation",
              "disabled_by_seller",
              "disabled_by_takealot",
              "conversion_percentage_30_days",
              "conversion_percentage_previous_30_days",
              "page_views_30_days",
              "quantity_returned_30_days",
              "condition",
              "image_url",
              "productline_id",
              "width_cm",
              "length_cm",
              "height_cm",
              "weight_grams",
              "is_conveyable",
              "benchmark_price",
              "total_wishlist",
              "wishlist_30_days",
              "listing_quality"
            ]
          }
        }
      },
      "OfferExpands": {
        "description": "Expand the response with related entities.\n\n<b>Note on</b> `replenishment_blocks`: Including this enum will only return a response\nif the offer is currently subject to an explicit, full replenishment block. It\ndoes not return data regarding replenishment limits (including instances where\nthe limit is 0, which acts as a block) or replenishment allows. If the offer\nis not fully blocked, no replenishment data will be returned.\n",
        "name": "expands",
        "in": "query",
        "required": false,
        "example": [
          "seller_warehouse_stock",
          "takealot_warehouse_stock"
        ],
        "schema": {
          "type": "array",
          "items": {
            "type": "string",
            "enum": [
              "seller_warehouse_stock",
              "offer_charges",
              "takealot_warehouse_stock",
              "replenishment_blocks"
            ]
          }
        }
      },
      "Limit": {
        "description": "Limit the number of results returned. Default is 100, maximum is 1000.",
        "name": "limit",
        "in": "query",
        "required": false,
        "example": 50,
        "schema": {
          "type": "integer",
          "default": 100,
          "minimum": 1,
          "maximum": 1000
        }
      },
      "ContinuationToken": {
        "description": "Continuation token to get the next page of results. Obtained from a previous response.",
        "name": "continuation_token",
        "in": "query",
        "required": false,
        "example": "eyJvZmZzZXQiOiAxMDB9",
        "schema": {
          "type": "string",
          "minimum": 1
        }
      },
      "IncludeCount": {
        "description": "Include a count of total items in the response.",
        "name": "include_count",
        "in": "query",
        "required": false,
        "example": true,
        "schema": {
          "type": "boolean",
          "default": false
        }
      },
...
      "SaleFields": {
        "description": "Return only these fields.",
        "name": "fields",
        "in": "query",
        "required": false,
        "example": [
          "order_id",
          "order_date",
          "sale_status",
          "selling_price",
          "quantity"
        ],
        "schema": {
          "type": "array",
          "items": {
            "type": "string",
            "enum": [
              "order_item_id",
              "order_id",
              "order_date",
              "sale_status",
              "offer_id",
              "tsin_id",
              "sku",
              "selling_price",
              "quantity",
              "success_fee",
              "fulfillment_fee",
              "courier_collection_fee",
              "total_fees",
              "stock_transfer_fee",
              "sales_region",
              "stock_source_region"
            ]
          }
        }
      },
...
    },
    "responses": {
      "DefaultErrorResponse": {
        "description": "General response format for non-200 responses",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/DefaultError"
            }
          }
        }
      },
      "ListOffersResponse": {
        "description": "Response payload for listing offers",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "description": "A paginated list of offers.",
              "properties": {
                "items": {
                  "type": "array",
                  "description": "The list of offers in this page of results.",
                  "items": {
                    "$ref": "#/components/schemas/Offer"
                  }
                },
                "count": {
                  "$ref": "#/components/schemas/Count"
                },
                "limit": {
                  "$ref": "#/components/schemas/Limit"
                },
                "continuation_token": {
                  "$ref": "#/components/schemas/ContinuationToken"
                }
              },
              "additionalProperties": false
            }
          }
        }
      },
      "ForbiddenErrorResponse": {
        "description": "API key is missing or invalid",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ForbiddenError"
            }
          }
        }
      },
      "ValidationErrorResponse": {
        "description": "HTTP 400 errors with sub-types",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ValidationError"
            }
          }
        }
      },
      "ErrorResponse": {
        "description": "HTTP 500 errors with sub-types",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Error"
            }
          }
        }
      },
...
      "ListSalesResponse": {
        "description": "Response payload for listing sales",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "description": "A paginated list of sales.",
              "properties": {
                "items": {
                  "type": "array",
                  "description": "The list of sales in this page of results.",
                  "items": {
                    "$ref": "#/components/schemas/Sale"
                  }
                },
                "count": {
                  "$ref": "#/components/schemas/Count"
                },
                "limit": {
                  "$ref": "#/components/schemas/Limit"
                },
                "continuation_token": {
                  "$ref": "#/components/schemas/ContinuationToken"
                }
              },
              "additionalProperties": false
            }
          }
        }
      },
...
    },
    "requestBodies": {
      "CreateOfferRequest": {
        "description": "Request body for creating an offer.",
        "required": true,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "minProperties": 1,
              "additionalProperties": false,
              "properties": {
                "barcode": {
                  "type": "string",
                  "minLength": 1,
                  "required": true,
                  "description": "Barcode for the offer (provided by the merchant)",
                  "example": "MPTALX12345678-9"
                },
                "sku": {
                  "$ref": "#/components/schemas/SKURequestBody"
                },
                "minimum_leadtime_days": {
                  "type": "integer",
                  "description": "The leadtime in days for the offer to be shipped to a Takealot facility when shipped from the seller's warehouse.",
                  "example": 5
                },
                "selling_price": {
                  "$ref": "#/components/schemas/SellingPriceRequestBody"
                },
                "rrp": {
                  "$ref": "#/components/schemas/RRPRequestBody"
                },
                "seller_warehouse_stock": {
                  "type": "array",
                  "description": "Stock levels in the seller's warehouses.",
                  "items": {
                    "type": "object",
                    "description": "Stock quantity at a specific seller warehouse.",
                    "required": [
                      "seller_warehouse_id",
                      "quantity_available"
                    ],
                    "properties": {
                      "seller_warehouse_id": {
                        "type": "integer",
                        "description": "The ID of the seller's warehouse. Assigned by Takealot.",
                        "example": 1234
                      },
                      "quantity_available": {
                        "type": "integer",
                        "description": "The quantity of stock that is available for purchase from this warehouse.",
                        "example": 100
                      }
                    }
                  }
                }
              }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/offers": {
      "get": {
        "summary": "List Offers",
        "description": "List Offers.",
        "operationId": "resources.offers.Offers.on_get",
        "parameters": [
          {
            "$ref": "#/components/parameters/OfferID"
          },
          {
            "$ref": "#/components/parameters/OfferIDIn"
          },
          {
            "name": "status",
            "description": "Only include offers with this status.",
            "in": "query",
            "required": false,
            "example": "buyable",
            "schema": {
              "type": "string",
              "enum": [
                "buyable",
                "not_buyable",
                "disabled_by_seller",
                "disabled_by_takealot"
              ]
            }
          },
          {
            "name": "offer_id__gt",
            "description": "Only include offers with IDs greater than this value.",
            "in": "query",
            "required": false,
            "example": 1000000,
            "schema": {
              "$ref": "#/components/schemas/OfferID"
            }
          },
          {
            "name": "offer_id__lt",
            "description": "Only include offers with IDs less than this value.",
            "in": "query",
            "required": false,
            "example": 2000000,
            "schema": {
              "$ref": "#/components/schemas/OfferID"
            }
          },
          {
            "name": "sku",
            "description": "Only include offers with this SKU.",
            "in": "query",
            "required": false,
            "example": "SKU-ABC123",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "barcode",
            "description": "Only include offers with this barcode.",
            "in": "query",
            "required": false,
            "example": "MPTALX12345678-9",
            "schema": {
              "type": "string"
            }
          },
          {
            "$ref": "#/components/parameters/OfferFields"
          },
          {
            "$ref": "#/components/parameters/OfferExpands"
          },
          {
            "$ref": "#/components/parameters/Limit"
          },
          {
            "$ref": "#/components/parameters/ContinuationToken"
          },
          {
            "name": "order_by",
            "description": "Order the results by these fields. Prefix with '-' for descending order.",
            "in": "query",
            "required": false,
            "example": "-offer_id",
            "schema": {
              "type": "string",
              "minimum": 1,
              "enum": [
                "offer_id",
                "-offer_id"
              ],
              "default": "offer_id"
            }
          },
          {
            "$ref": "#/components/parameters/IncludeCount"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ListOffersResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenErrorResponse"
          },
          "4XX": {
            "$ref": "#/components/responses/ValidationErrorResponse"
          },
          "5XX": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/DefaultErrorResponse"
          }
        },
        "tags": [
          "Offers"
        ]
      },
      "post": {
        "summary": "Create offer",
        "description": "Create a new offer.",
        "operationId": "resources.offers.Offers.on_post",
        "parameters": [
          {
            "$ref": "#/components/parameters/OfferFields"
          },
          {
            "$ref": "#/components/parameters/OfferExpands"
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/CreateOfferRequest"
        },
        "responses": {
          "201": {
            "description": "Offer created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Offer"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenErrorResponse"
          },
          "4XX": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "5XX": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/DefaultErrorResponse"
          }
        },
        "tags": [
          "Offers"
        ]
      }
    },
...
    "/sales": {
      "get": {
        "summary": "List Sales",
        "description": "List all your sales.",
        "operationId": "resources.sales.Sales.on_get",
        "parameters": [
          {
            "name": "order_id",
            "description": "Filter sales by order ID.",
            "in": "query",
            "required": false,
            "example": 9876543,
            "schema": {
              "type": "integer",
              "minimum": 1
            }
          },
          {
            "name": "order_date__lte",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": false,
            "example": "2025-12-31",
            "description": "Paired with order_date__gte, together forming an inclusive daterange for order_date, in SAST."
          },
          {
            "name": "order_date__gte",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            },
            "required": false,
            "example": "2025-01-01",
            "description": "Paired with order_date__lte, together forming an inclusive daterange for order_date, in SAST."
          },
          {
            "name": "stock_source_region",
            "description": "Filter sales by the region from which the stock was sourced.",
            "in": "query",
            "required": false,
            "example": "JHB",
            "schema": {
              "type": "string",
              "enum": [
                "cpt",
                "CPT",
                "jhb",
                "JHB",
                "dbn",
                "DBN"
              ]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "example": 50,
            "schema": {
              "type": "integer",
              "default": 20,
              "minimum": 1,
              "maximum": 100
            },
            "description": "The number of sales per page."
          },
          {
            "$ref": "#/components/parameters/SaleFields"
          },
          {
            "$ref": "#/components/parameters/ContinuationToken"
          },
          {
            "$ref": "#/components/parameters/IncludeCount"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/ListSalesResponse"
          },
          "403": {
            "$ref": "#/components/responses/ForbiddenErrorResponse"
          },
          "4XX": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "5XX": {
            "$ref": "#/components/responses/ErrorResponse"
          },
          "default": {
            "$ref": "#/components/responses/DefaultErrorResponse"
          }
        },
        "tags": [
          "Sales"
        ]
      }
    },
...
  }
}